1876 Improved support for MZ executables - Fixed incorrect header - Gracefully handle segment overflows - Added missing segment relocations - Create uninitiallized memory according to header
